Refactor Header to accomodate the new architecture
