+++ packages/bierzo-wallet/src/logic/transactions/types/BwSendTransaction/ui/SendTxHeader/index.tsx
@@ -6,17 +6,17 @@ import Hairline from 'medulas-react-components/lib/components/Hairline';
 import Image from 'medulas-react-components/lib/components/Image';
 import * as React from 'react';
 
-import { history } from '../../../../../routes';
-import { PAYMENT_ROUTE } from '../../../../../routes/paths';
-import { ProcessedTx } from '../../../../../store/notifications';
-import { itemBackground } from '../../../../../theme/css';
-import { prettyAmount } from '../../../../../utils/balances';
-import receiveTx from '../../../assets/transactionReceive.svg';
-import sendTx from '../../../assets/transactionSend.svg';
+import { BwSendProps } from '../..';
+import { history } from '../../../../../../routes';
+import { PAYMENT_ROUTE } from '../../../../../../routes/paths';
+import { itemBackground } from '../../../../../../theme/css';
+import { prettyAmount } from '../../../../../../utils/balances';
+import receiveTx from '../assets/transactionReceive.svg';
+import sendTx from '../assets/transactionSend.svg';
 import Msg from './MsgSuccess';
 
 interface ItemProps {
-  readonly item: ProcessedTx;
+  readonly item: BwSendProps;
   readonly lastOne: boolean;
 }
